Add docker-compose configuration to repository.
Add an initial docker-compose.yml. Still seems pretty fragile due to the
inherent nature of containers having a strong bias towards web frameworks.

When starting the Asterisk portion and mounting the volumes, things work fine
but you'll never have access to the console. You need to use a combination of
the `docker-compose up` to start the volumes, and the `docker-compose run` to
start up the Asterisk container.

Hopefully in the future things will be slight less web biased and more
generally microservices based.